% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell.R
\name{shell}
\alias{shell}
\title{there are no functions in this package at this time, just enjoy the vignettes}
\usage{
shell()
}
\description{
there are no functions in this package at this time, just enjoy the vignettes
}
